Tweak: Screentips/Say fonts
On the local server I checked whisper/yell/emotion/say, but could possibly have lost a letter from the font when editing or made the wrong indentation, which could letter smudge. If it's going to merge, you need to check at least a one real round. And if something is broken roll it back for revision.